requirements.txt
argparse~=1.4.0
logging~=0.4.9.6
PyYAML~=6.0.2
urllib3~=2.3.0
requests~=2.32.3